// src/app/api/participants/model/participantDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * ParticipantDto to be saved
 */
export interface ParticipantDto {
  id: string;
  name: string;
  email: string;
  phone: string;
  dormId?: string;
  dormName: string;
  addressId?: string;
  addressLine1: string;
  addressLine2?: string;
  addressCity: string;
  addressPostalCode: string;
}